only otelException is exported
